+++ src/main.cpp
@@ -1,7 +1,7 @@
 // Wisp plug-in
 #include "miscellaneous/settings.hpp"
 #include "plugin/plugin_main.hpp"
-#include "plugin/overrides/viewport_renderer_override.hpp"
+#include "plugin/viewport_renderer.hpp"
 
 // Maya API
 #include <maya/MFnPlugin.h>
