+++ src/plugin/parsers/scene_graph_parser.cpp
