+++ src/plugin/parsers/scene_graph_parser.hpp
