+++ src/plugin/render_operations/quad_renderer_override.hpp
